# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cash": {
      "post": {
        "operationId": "CashController_addCash",
        "summary": "Withdrawal Request [24130009- 複委託出金申請(驗章)]",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCashDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CashResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "cash"
        ]
      },
      "get": {
        "operationId": "CashController_queryCash",
        "summary": "Withdrawal Query [50004002- OLTSAPI0047]",
        "parameters": [
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20001365441",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "CURRENCY",
            "required": true,
            "in": "query",
            "description": "Currency",
            "schema": {
              "example": "USD",
              "enum": [
                "USD",
                "HKD",
                "JPY",
                "THB",
                "SGD",
                "KRW",
                "EUR",
                "GBP",
                "CNY",
                "AUD",
                "NZD",
                "ZAR"
              ],
              "type": "string"
            }
          },
          {
            "name": "APPLY_DATE",
            "required": true,
            "in": "query",
            "description": "Application Date",
            "schema": {
              "example": "2021/10/10",
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CashResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "cash"
        ]
      }
    },
    "/cash/delete": {
      "post": {
        "operationId": "CashController_deleteCash",
        "summary": "Withdrawal Delete [2413000A- 複委託出金刪除(驗章)]",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCashDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CashResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "cash"
        ]
      }
    },
    "/cash/detail": {
      "get": {
        "operationId": "CashController_getCashReport",
        "summary": "Withdrawal Detail [50004006- OLTSAPI0049]",
        "parameters": [
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20001365441",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "CURRENCY",
            "required": false,
            "in": "query",
            "description": "Currency",
            "schema": {
              "example": "USD",
              "enum": [
                "USD",
                "HKD",
                "JPY",
                "THB",
                "SGD",
                "KRW",
                "EUR",
                "GBP",
                "CNY",
                "AUD",
                "NZD",
                "ZAR"
              ],
              "type": "string"
            }
          },
          {
            "name": "APPLY_START_DATE",
            "required": true,
            "in": "query",
            "description": "Application Date(From)",
            "schema": {
              "example": "2021/10/10",
              "type": "string"
            }
          },
          {
            "name": "APPLY_END_DATE",
            "required": true,
            "in": "query",
            "description": "Application Date(to)",
            "schema": {
              "example": "2021/10/11",
              "type": "string"
            }
          },
          {
            "name": "PHASE",
            "required": true,
            "in": "query",
            "description": "Withdrawal Phase",
            "schema": {
              "example": "0",
              "enum": [
                "0",
                "1",
                "2"
              ],
              "type": "string"
            }
          },
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CashReportResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "cash"
        ]
      }
    },
    "/cash/unhold-amount": {
      "get": {
        "operationId": "CashController_getUnHoldAmount",
        "summary": "Unhold Request [24130008- 複委託解圈申請(驗章)]",
        "parameters": [
          {
            "name": "dtaSrc",
            "required": true,
            "in": "query",
            "description": "平台來源別",
            "schema": {
              "example": "o",
              "type": "string"
            }
          },
          {
            "name": "IP",
            "required": true,
            "in": "query",
            "description": "Customer IP",
            "schema": {
              "example": "00.00.00.00",
              "type": "string"
            }
          },
          {
            "name": "CUSTOMER_ID",
            "required": true,
            "in": "query",
            "description": "Customer account",
            "schema": {
              "example": "20001365441",
              "type": "string"
            }
          },
          {
            "name": "ID",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "L220289830",
              "type": "string"
            }
          },
          {
            "name": "CURRENCY",
            "required": true,
            "in": "query",
            "description": "Currency",
            "schema": {
              "example": "USD",
              "enum": [
                "USD",
                "HKD",
                "JPY",
                "THB",
                "SGD",
                "KRW",
                "EUR",
                "GBP",
                "CNY",
                "AUD",
                "NZD",
                "ZAR"
              ],
              "type": "string"
            }
          },
          {
            "name": "APPLY_DATE",
            "required": true,
            "in": "query",
            "description": "Application Date",
            "schema": {
              "example": "2022/10/27",
              "type": "string"
            }
          },
          {
            "name": "SYSTEM_ID",
            "required": false,
            "in": "query",
            "description": "Order placement channel",
            "schema": {
              "example": "RU",
              "enum": [
                "AG",
                "SV",
                "CH",
                "AP",
                "MJ",
                "XT",
                "CW",
                "BT",
                "MM",
                "WM",
                "ES",
                "H5",
                "RU"
              ],
              "type": "string"
            }
          },
          {
            "name": "caCert",
            "required": true,
            "in": "query",
            "description": "CA Cert(憑證序號)",
            "schema": {
              "example": "12345678",
              "type": "string"
            }
          },
          {
            "name": "caLen",
            "required": false,
            "in": "query",
            "description": "CA密文長度",
            "schema": {
              "example": "0",
              "type": "string"
            }
          },
          {
            "name": "encodedCA",
            "required": true,
            "in": "query",
            "description": "CA密文",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "orderData",
            "required": false,
            "in": "query",
            "description": "",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "Market api token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "responseTime": {
                          "type": "string",
                          "example": "2024-01-01 23:59:59"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UnholdAMountResponse"
                          }
                        },
                        "errors": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "cash"
        ]
      }
    },
...
  },
  "info": {
    "title": "KGI Securities API spec",
    "description": "Internal APIs for KGI",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3057/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "scheme": "apikey"
      }
    },
    "schemas": {
...
      "CashResponse": {
        "type": "object",
        "properties": {
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441"
          },
          "CUSTOMER_NAME": {
            "type": "string",
            "example": "米奇"
          },
          "APPLY_DATE": {
            "type": "string",
            "example": "2022/10/27"
          },
          "CURRENCY": {
            "type": "string",
            "example": "HKD"
          },
          "TODAY_DROW": {
            "type": "number",
            "example": 2461
          },
          "TODAY_DEPOSIT": {
            "type": "number",
            "example": 0
          },
          "TRANS": {
            "type": "number",
            "example": 0.41
          },
          "BANK": {
            "type": "string",
            "example": "外幣自有帳戶"
          },
          "ACTION_TYPE": {
            "type": "string",
            "example": "一般"
          }
        },
        "required": [
          "CUSTOMER_ID",
          "CUSTOMER_NAME",
          "APPLY_DATE",
          "CURRENCY",
          "TODAY_DROW",
          "TODAY_DEPOSIT",
          "TRANS",
          "BANK",
          "ACTION_TYPE"
        ]
      },
      "AddCashDto": {
        "type": "object",
        "properties": {
          "dtaSrc": {
            "type": "string",
            "example": "o",
            "description": "dtaSrc"
          },
          "IP": {
            "type": "string",
            "example": "00.00.00.00",
            "description": "Customer IP"
          },
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441",
            "description": "Customer account"
          },
          "ID": {
            "type": "string",
            "example": "L220289830",
            "description": "Customer ID"
          },
          "CURRENCY": {
            "type": "string",
            "enum": [
              "USD",
              "HKD",
              "JPY",
              "THB",
              "SGD",
              "KRW",
              "EUR",
              "GBP",
              "CNY",
              "AUD",
              "NZD",
              "ZAR"
            ],
            "example": "USD",
            "description": "Currency"
          },
          "AMOUNT": {
            "type": "number",
            "example": 0,
            "description": "Commission Amount"
          },
          "APPLY_DATE": {
            "type": "string",
            "example": "2021/10/10",
            "description": "Application Date"
          },
          "SYSTEM_ID": {
            "type": "string",
            "enum": [
              "AG",
              "SV",
              "CH",
              "AP",
              "MJ",
              "XT",
              "CW",
              "BT",
              "MM",
              "WM",
              "ES",
              "H5",
              "RU"
            ],
            "example": "RU",
            "description": "Order placement channel"
          },
          "caCert": {
            "type": "string",
            "example": "8",
            "description": "CA Cert(憑證序號)"
          },
          "caLen": {
            "type": "number",
            "example": 0,
            "description": "CA密文長度"
          },
          "encodedCA": {
            "type": "string",
            "example": "caLen8",
            "description": "CA密文"
          },
          "orderData": {
            "type": "string",
            "example": "",
            "description": ""
          }
        },
        "required": [
          "IP",
          "CUSTOMER_ID",
          "ID",
          "CURRENCY",
          "AMOUNT",
          "APPLY_DATE",
          "caCert",
          "encodedCA"
        ]
      },
      "DeleteCashDto": {
        "type": "object",
        "properties": {
          "dtaSrc": {
            "type": "string",
            "example": "o",
            "description": "dtaSrc"
          },
          "IP": {
            "type": "string",
            "example": "00.00.00.00",
            "description": "Customer IP"
          },
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441",
            "description": "Customer account"
          },
          "ID": {
            "type": "string",
            "example": "L220289830",
            "description": "Customer ID"
          },
          "CURRENCY": {
            "type": "string",
            "enum": [
              "USD",
              "HKD",
              "JPY",
              "THB",
              "SGD",
              "KRW",
              "EUR",
              "GBP",
              "CNY",
              "AUD",
              "NZD",
              "ZAR"
            ],
            "example": "USD",
            "description": "Currency"
          },
          "APPLY_DATE": {
            "type": "string",
            "example": "2021/10/10",
            "description": "Application Date"
          },
          "SYSTEM_ID": {
            "type": "string",
            "enum": [
              "AG",
              "SV",
              "CH",
              "AP",
              "MJ",
              "XT",
              "CW",
              "BT",
              "MM",
              "WM",
              "ES",
              "H5",
              "RU"
            ],
            "example": "RU",
            "description": "Order placement channel"
          },
          "caCert": {
            "type": "string",
            "example": "8",
            "description": "CA Cert(憑證序號)"
          },
          "caLen": {
            "type": "number",
            "example": 4,
            "description": "CA密文長度"
          },
          "encodedCA": {
            "type": "string",
            "example": "caLen8",
            "description": "CA密文"
          },
          "orderData": {
            "type": "string",
            "example": "",
            "description": ""
          }
        },
        "required": [
          "IP",
          "CUSTOMER_ID",
          "ID",
          "CURRENCY",
          "APPLY_DATE",
          "caCert",
          "caLen",
          "encodedCA"
        ]
      },
      "CashReportResponse": {
        "type": "object",
        "properties": {
          "BRANCH_NAME": {
            "type": "string",
            "example": "台北分公司"
          },
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441"
          },
          "CUSTOMER_NAME": {
            "type": "string",
            "example": "米奇"
          },
          "APPLY_DATE": {
            "type": "string",
            "example": "2022/10/27"
          },
          "CURRENCY": {
            "type": "string",
            "example": "HKD"
          },
          "AMOUNT": {
            "type": "number",
            "example": 2461
          },
          "CHANNEL": {
            "type": "string",
            "example": "RU"
          }
        },
        "required": [
          "BRANCH_NAME",
          "CUSTOMER_ID",
          "CUSTOMER_NAME",
          "APPLY_DATE",
          "CURRENCY",
          "AMOUNT",
          "CHANNEL"
        ]
      },
      "UnholdAMountResponse": {
        "type": "object",
        "properties": {
          "BRANCH_NAME": {
            "type": "string",
            "example": "台北分公司"
          },
          "CUSTOMER_ID": {
            "type": "string",
            "example": "20001365441"
          },
          "CUSTOMER_NAME": {
            "type": "string",
            "example": "米奇"
          },
          "APPLY_DATE": {
            "type": "string",
            "example": "2022/10/27"
          },
          "CURRENCY": {
            "type": "string",
            "example": "HKD"
          },
          "AMOUNT": {
            "type": "number",
            "example": 2461
          }
        },
        "required": [
          "BRANCH_NAME",
          "CUSTOMER_ID",
          "CUSTOMER_NAME",
          "APPLY_DATE",
          "CURRENCY",
          "AMOUNT"
        ]
      },
...
    }
  }
}